seguridad en los servicios y token modificado

+++ src/main/java/api_incidencias/api_incidencias/Servicios/ClienteService.java
@@ -15,7 +15,8 @@ import java.util.Optional;
 public class ClienteService {
     @Autowired
     private RepositorioCliente reposCliente;
-
+    @Autowired
+    private Seguridad seguridad;
     public Cliente addCliente(Cliente cliente) {
         return reposCliente.save(cliente);
     }
@@ -40,37 +41,40 @@ public class ClienteService {
      * @return
      */
     public Cliente updateCliente(Long idCliente, Cliente cliente) {
-        Optional<Cliente> clienteExistenteOptional = reposCliente.findById(idCliente);
+        if (seguridad.isAdmin() || seguridad.isElMismo(cliente.getCorreoElectronico())){
+            Optional<Cliente> clienteExistenteOptional = reposCliente.findById(idCliente);
 
-        if (clienteExistenteOptional.isPresent()) {
-            Cliente clienteExixtente = clienteExistenteOptional.get();
+            if (clienteExistenteOptional.isPresent()) {
+                Cliente clienteExixtente = clienteExistenteOptional.get();
 
-            if (idCliente.equals(cliente.getIdUsuario())) {
-                // Actualizo los atributos del libro existente con los del libro proporcionado
-                clienteExixtente.setDni(cliente.getDni());
-                clienteExixtente.setNombre(cliente.getNombre());
-                clienteExixtente.setApellido(cliente.getApellido());
-                clienteExixtente.setCorreoElectronico(cliente.getCorreoElectronico());
-                clienteExixtente.setContrasena(cliente.getContrasena());
+                if (idCliente.equals(cliente.getIdUsuario())) {
+                    // Actualizo los atributos del libro existente con los del libro proporcionado
+                    clienteExixtente.setDni(cliente.getDni());
+                    clienteExixtente.setNombre(cliente.getNombre());
+                    clienteExixtente.setApellido(cliente.getApellido());
+                    clienteExixtente.setCorreoElectronico(cliente.getCorreoElectronico());
+                    clienteExixtente.setContrasena(cliente.getContrasena());
 
-                clienteExixtente.setFechaRegistro(cliente.getFechaRegistro());
-                clienteExixtente.setImagenPerfil(cliente.getImagenPerfil());
-                clienteExixtente.setTelefono(cliente.getTelefono());
+                    clienteExixtente.setFechaRegistro(cliente.getFechaRegistro());
+                    clienteExixtente.setImagenPerfil(cliente.getImagenPerfil());
+                    clienteExixtente.setTelefono(cliente.getTelefono());
 
-                clienteExixtente.setCalle(cliente.getCalle());
-                clienteExixtente.setCiudad(cliente.getCiudad());
-                clienteExixtente.setProvincia(cliente.getProvincia());
-                clienteExixtente.setCodigoPostal(cliente.getCodigoPostal());
-                clienteExixtente.setPais(cliente.getPais());
+                    clienteExixtente.setCalle(cliente.getCalle());
+                    clienteExixtente.setCiudad(cliente.getCiudad());
+                    clienteExixtente.setProvincia(cliente.getProvincia());
+                    clienteExixtente.setCodigoPostal(cliente.getCodigoPostal());
+                    clienteExixtente.setPais(cliente.getPais());
 
-                // Guarda el usuario actualizado en el repositorio
-                return reposCliente.save(clienteExixtente);
+                    // Guarda el usuario actualizado en el repositorio
+                    return reposCliente.save(clienteExixtente);
+                } else {
+                    throw new IllegalArgumentException("El id proporcionado no coincide con el ID del usuario.");
+                }
             } else {
-                throw new IllegalArgumentException("El id proporcionado no coincide con el ID del usuario.");
+                throw new IllegalArgumentException("El usuario con el ID proporcionado no existe.");
             }
-        } else {
-            throw new IllegalArgumentException("El usuario con el ID proporcionado no existe.");
         }
+        throw new IllegalArgumentException("No tienes permisos");
     }
 
     /**
@@ -80,18 +84,20 @@ public class ClienteService {
      * @return
      */
     public ResponseEntity<String> deleteCliente(Long id) {
+        if (seguridad.isAdmin()){
+            Optional<Cliente> userOptional = reposCliente.findById(id);
+            if (userOptional.isPresent()) {
+                reposCliente.deleteById(id);
 
-        Optional<Cliente> userOptional = reposCliente.findById(id);
-        if (userOptional.isPresent()) {
-            reposCliente.deleteById(id);
-            ;
-
-            return ResponseEntity.status(HttpStatus.NO_CONTENT)
-                    .body("Usuario eliminado correctamente.");
-        } else {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .body("No se encontró el usuario correspondiente.");
+                return ResponseEntity.status(HttpStatus.NO_CONTENT)
+                        .body("Usuario eliminado correctamente.");
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                        .body("No se encontró el usuario correspondiente.");
+            }
         }
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                .body("No tienes permisos.");
     }
 
 }
